全ての表記をlong型に統一

+++ src/com/github/unchama/gui/ranking/fishing/WeekFishingExpRankingMenuManager.java
@@ -14,7 +14,7 @@ import com.github.unchama.sql.moduler.RankingTableManager.TimeType;
 public class WeekFishingExpRankingMenuManager extends RankingMenuManager {
 	@Override
 	protected String getLore(double value) {
-		return "総釣り経験値:" + (int)value;
+		return "総釣り経験値:" + (long)value;
 	}
 
 	@Override
