+++ src/com/github/unchama/gui/ranking/huntingexp/TotalHuntingExpRankingMenuManager.java
@@ -14,7 +14,7 @@ public class TotalHuntingExpRankingMenuManager extends TotalRankingMenuManager {
 
 	@Override
 	protected String getLore(double value) {
-		return "総狩猟経験値:" + (int)value;
+		return "総狩猟経験値:" + (long)value;
 	}
 
 	@Override
